+++ R/reexports.R
@@ -1,53 +1,7 @@
-# ----------------------- tidyverse -------------------------------------
-
 #' @export
-#' @importFrom statsExpressions "%>%"
-statsExpressions::`%>%`
-
-#' @export
-#' @importFrom statsExpressions "%<>%"
-statsExpressions::`%<>%`
-
-#' @export
-#' @importFrom statsExpressions "%$%"
-statsExpressions::`%$%`
-
-#' @export
-#' @importFrom statsExpressions "%<-%"
-statsExpressions::`%<-%`
-
-#' @export
-#' @importFrom statsExpressions tibble
-statsExpressions::tibble
+#' @importFrom magrittr "%>%"
+magrittr::`%>%`
 
 #' @export
 #' @importFrom statsExpressions pairwise_comparisons
 statsExpressions::pairwise_comparisons
-
-#' @export
-#' @importFrom statsExpressions as_tibble
-statsExpressions::as_tibble
-
-#' @export
-#' @importFrom rlang exec
-rlang::exec
-
-#' @export
-#' @importFrom rlang "!!"
-rlang::`!!`
-
-#' @export
-#' @importFrom rlang "!!!"
-rlang::`!!!`
-
-#' @export
-#' @importFrom rlang "%|%"
-rlang::`%|%`
-
-#' @export
-#' @importFrom rlang "%||%"
-rlang::`%||%`
-
-#' @export
-#' @importFrom rlang ":="
-rlang::`:=`