+++ tests/testthat.R
@@ -1,5 +1,10 @@
 library(testthat)
 library(ggstatsplot)
-library(statsExpressions)
+
+# suppress printing environment name (noisy)
+invisible({
+  loadNamespace("statsExpressions")
+  loadNamespace("vdiffr")
+})
 
 test_check("ggstatsplot")